+++ packages/ui-components/src/features/Auth/ResetPasswordForm/ResetPasswordForm.tsx
@@ -1,0 +1,61 @@
+/*
+ * Tupaia
+ * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+import React from 'react';
+import styled from 'styled-components';
+import { LinkProps } from 'react-router-dom';
+import { SubmitHandler, useForm } from 'react-hook-form';
+import { AuthErrorMessage } from '../AuthErrorMessage';
+import { PasswordForm } from './PasswordForm';
+import { ResetPasswordSuccess } from './ResetPasswordSuccess';
+
+const Wrapper = styled.div`
+  width: 36rem;
+  max-width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  > div {
+    width: 100%;
+  }
+`;
+
+interface ResetPasswordFormProps {
+  isLoading: boolean;
+  isSuccess?: boolean;
+  error?: Error | null;
+  formContext: ReturnType<typeof useForm>;
+  onSubmit: SubmitHandler<any>;
+  passwordResetToken?: string | null;
+  loginLink: LinkProps['to'];
+}
+
+export const ResetPasswordForm = ({
+  passwordResetToken,
+  onSubmit,
+  isSuccess,
+  error,
+  isLoading,
+  formContext,
+  loginLink,
+}: ResetPasswordFormProps) => {
+  return (
+    <Wrapper>
+      {error && <AuthErrorMessage>{error.message}</AuthErrorMessage>}
+      {isSuccess ? (
+        <ResetPasswordSuccess loginLink={loginLink} />
+      ) : (
+        <PasswordForm
+          passwordResetToken={passwordResetToken}
+          onSubmit={onSubmit}
+          isLoading={isLoading}
+          formContext={formContext}
+          loginLink={loginLink}
+        />
+      )}
+    </Wrapper>
+  );
+};
