+++ packages/ui-components/src/features/Auth/ResetPasswordForm/index.ts
@@ -1,0 +1,1 @@
+export { ResetPasswordForm } from './ResetPasswordForm';
